requirements.txt
asgiref==3.5.0
Django==3.2
django-cors-headers==3.11.0
djangorestframework==3.12.0
pytz==2021.3
sqlparse==0.4.2
